requirements.txt
mcp>=1.20
anthropic>=0.80
openpyxl>=3.1
python-dotenv>=1.0
pydantic>=2.0
